requirements.txt
aiohttp==3.6.3
astroid @ file:///opt/concourse/worker/volumes/live/21fd14a9-2a7e-484b-7394-5a9912cdcf80/volume/astroid_1592498459180/work
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
boto3==1.16.49
botocore==1.19.49
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
discord.py==1.5.1
idna==2.10
isort @ file:///tmp/build/80754af9/isort_1602603989581/work
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pycparser==2.20
pylint @ file:///opt/concourse/worker/volumes/live/ed0164b6-bcc7-4f6b-7dd4-ad89660b5dcb/volume/pylint_1598624018129/work
PyNaCl==1.4.0
python-dateutil==2.8.1
python-dotenv==0.15.0
s3transfer==0.3.3
six @ file:///opt/concourse/worker/volumes/live/5b31cb27-1e37-4ca5-6e9f-86246eb206d2/volume/six_1605205320872/work
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
urllib3==1.26.2
wavelink==0.9.6
wrapt==1.11.2
yarl==1.5.1
youtube-dl==2020.11.21.1
